refactor(transactions): extract shared OrderLineItemRow (SO/PO)
